SONAR-8857 move code from SearchDataLoader to SearchAction
